refactor: split admin views into dedicated pages and add meeting detail view
