+++ web/b3desk/static/css/global.css
@@ -547,3 +547,7 @@ This is a fix until we deal with the update.
 .fr-header__brand-top {
   flex-shrink: 0;
 }
+
+.fr-btn--delete.fr-btn--admin {
+  color: white !important;
+}
