--- web/b3desk/templates/admin/options.html
@@ -1,24 +1,0 @@
-<a class="fr-btn" href="{{ url_for("public.welcome") }}" title="{% trans %}Ouvrir la page d'administration{% endtrans %}">
-    {% trans %}Quitter la page d'administration{% endtrans %}
-</a>
-<br>
-<br>
-<div class="fr-py-3w main-connected-actions">
-    <div class="fr-container">
-        <div class="fr-grid-row fr-grid-row--gutters">
-            <div class="fr-col">
-                <h4>{% trans %}Liste des utilisateurs{% endtrans %}</h4>
-                <a class="fr-btn" href="{{ url_for("admin.manage_users") }}" title="{% trans %}Liste des utilisateurs{% endtrans %}">
-                    {% trans %}Utilisateurs{% endtrans %}
-                </a>
-            </div>
-            <div class="fr-col">
-                <h4>{% trans %}Liste des réunions{% endtrans %}</h4>
-                <a class="fr-btn" href="{{ url_for("admin.manage_meetings") }}" title="{% trans %}Liste des réunions{% endtrans %}">
-                    {% trans %}Réunions{% endtrans %}
-                </a>
-            </div>
-        </div>
-    </div>
-</div>
-<br>
