+++ web/b3desk/templates/admin/user_table.html
@@ -1,39 +1,37 @@
 <div class="fr-grid-row fr-grid-row--gutters">
     <div class="fr-col-12">
-        <div class="fr-table fr-table--no-caption" id="meetings-component">
+        <div class="fr-table fr-table--no-caption" id="users-component">
             <div class="fr-table__wrapper ">
                 <div class="fr-table__container">
                     <div class="fr-table__content">
-                        <table id="meetings">
+                        <table id="users">
                             <thead class="">
                                 <tr class="">
                                     <th class="fr-col fr-cell--center">{% trans %}Identifiant{% endtrans %}</th>
                                     <th class="fr-col fr-cell--center">{% trans %}Nom{% endtrans %}</th>
-                                    <th class="fr-col fr-cell--center">{% trans %}Visio-code{% endtrans %}</th>
+                                    <th class="fr-col fr-cell--center">{% trans %}Prénom{% endtrans %}</th>
+                                    <th class="fr-col fr-cell--center">{% trans %}Email{% endtrans %}</th>
                                     <th class="fr-col fr-cell--center">{% trans %}Consulter{% endtrans %}</th>
                                 </tr>
                             </thead>
                             <tbody class="">
-                                {% for meeting in meetings %}
+                                {% for user in users %}
                                     <tr class="">
                                         <td class="fr-col fr-cell--center">
-                                            {{ meeting.id }}
+                                            {{ user.id }}
                                         </td>
                                         <td class="fr-col fr-cell--center">
-                                            {{ meeting.name }}
+                                            {{ user.family_name }}
                                         </td>
                                         <td class="fr-col fr-cell--center">
-                                            {{ meeting.visio_code }}
+                                            {{ user.given_name }}
                                         </td>
-                                        {% if not meeting.is_shadow %}
-                                            <td class="fr-col fr-cell--center">
-                                                <a href="{{ url_for("admin.meeting_infos",  meeting=meeting) }}" rel="noopener" class="fr-btn fr-btn--secondary fr-icon-eye-line" title="{% trans meeting_name=meeting.name %}Éditer {{ meeting_name }}{% endtrans %}"></a>
-                                            </td>
-                                        {% else %}
-                                            <td class="fr-col fr-cell--center">
-                                                {% trans %}Réunion silencieuse (shadow_meeting){% endtrans %}
-                                            </td>
-                                        {% endif %}
+                                        <td class="fr-col fr-cell--center">
+                                            {{ user.email }}
+                                        </td>
+                                        <td class="fr-col fr-cell--center">
+                                            <a href="{{ url_for("admin.user_infos", user=user) }}" rel="noopener" class="fr-btn fr-btn--secondary fr-icon-eye-line" title="{% trans username=user.fullname %}{{ username }} infos{% endtrans %}"></a>
+                                        </td>
                                     </tr>
                                 {% endfor %}
                             </tbody>
