+++ web/b3desk/templates/tools.html
@@ -25,6 +25,13 @@
                 <a href="{{ account_management_url }}" class="fr-link fr-fi-lock-line fr-link--icon-left" target="_self">{{ g.user.fullname }}</a>
                 </li>
                 {% endif %} -->
+                {% if g.user.admin %}
+                    <li>
+                        <a class="fr-btn fr-btn--delete fr-btn--admin" href="{{ url_for("admin.home") }}" title="{% trans %}Ouvrir la page d'administration{% endtrans %}">
+                            {% trans %}Admin{% endtrans %}
+                        </a>
+                    </li>
+                {% endif %}
                 <li>
                     <a class="fr-btn" href="{{ url_for("public.logout") }}">
                         <div>
